requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-jazzmin==3.0.2
Pillow==12.1.1
psycopg2-binary==2.9.11
gunicorn==23.0.0
whitenoise[brotli]==6.9.0
dj-database-url==3.0.1
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
